requirements.txt
streamlit
openai-whisper
google-generativeai
textblob
torch
imageio-ffmpeg
faster-whisper
